refactor: move colors into separate module
