+++ src/colors.js
@@ -1,0 +1,6 @@
+export const colors = {
+  welcomeColor: "\x1b[34m",
+  promptColor: "\x1b[32m",
+  unknownCommandColor: "\x1b[31m",
+  resetColor: "\x1b[0m",
+};
